backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-security
flask-restful
flask-mail
celery
redis
python-dotenv
google-generativeai
python-dateutil
python-dateutil
google-generativeai
Flask-Mail
celery
redis
Flask-Security
Flask-Migrate
Flask-SQLAlchemy
Flask-RESTful
python-dotenv